move all the stuff out of src pkg
